amount can be decimal in inv. transactions. (115297) The factor between inventory units can be decimal . (1076560)

+++ com/turquaz/inventory/bl/InvBLCardAdd.java
@@ -76,7 +76,7 @@ public class InvBLCardAdd {
 
 	}
 
-	public void registerUnits(Integer cardId, Object unitObj, int factor)
+	public void registerUnits(Integer cardId, Object unitObj, BigDecimal factor)
 			throws Exception {
 
 		TurqInventoryCardUnit cardUnit = new TurqInventoryCardUnit();
